PC-微信公众号登录

+++ app/Jobs/CheckOrderStatus.php
@@ -35,14 +35,35 @@ class CheckOrderStatus implements ShouldQueue
         $result = $api->getOrder($this->order->api_orderid);
         if($result->code == 200) {
             $file = $api->downloadReport($this->order->api_orderid);
+            if($file) {
+                $path = 'downloads/report-' . $this->order->api_orderid . '.zip';
+                \Storage::delete($path);
+                \Storage::put($path, $file);
+                //存储pdf
+                $content = $api->extractReportPdf($this->order->api_orderid);
+                file_put_contents(public_path('/pdf/') . $this->order->orderid . '.pdf', $content);
+                $report_pdf_path = config('app.url') . '/pdf/' . $this->order->orderid . '.pdf';
+                \DB::transaction(function() use ($path, $result, $report_pdf_path) {
+                    $this->order->update([
+                        'report_path' => $path,
+                        'report_pdf_path' => $report_pdf_path,
+                        'rate' => str_replace('%', '', $result->data->orderCheck->apiResultSemblance),
+                    ]);
+                    if($this->order->status == 4) {
+                        dispatch(new OrderCheckedMsg($this->order));
+                    }
+                    try {
+                        $report = $api->extractReportDetail($this->order->api_orderid);
+                        $content = $report->data->content;
+                    } catch (\Exception $e) {
+                        $content = '';
+                    }
+                    $this->order->report()->create([
+                        'content' => $content
+                    ]);
+                });
+            }
 
-            $path = 'downloads/report-' . $this->order->api_orderid . '.zip';
-            \Storage::delete($path);
-            \Storage::put($path, $file);
-            //存储pdf
-            $content = $api->extractReportPdf($this->order->api_orderid);
-            file_put_contents(public_path('/pdf/') . $this->order->orderid . '.pdf', $content);
-            $report_pdf_path = config('app.url') . '/pdf/' . $this->order->orderid . '.pdf';
 
 //            info(storage_path('app/' . $path));
 //            //解压zip文件
@@ -77,25 +98,7 @@ class CheckOrderStatus implements ShouldQueue
 //                $zip->close();
 //            }
 
-            \DB::transaction(function() use ($path, $result, $report_pdf_path) {
-                $this->order->update([
-                    'report_path' => $path,
-                    'report_pdf_path' => $report_pdf_path,
-                    'rate' => str_replace('%', '', $result->data->orderCheck->apiResultSemblance),
-                ]);
-                if($this->order->status == 4) {
-                    dispatch(new OrderCheckedMsg($this->order));
-                }
-                try {
-                    $report = $api->extractReportDetail($this->order->api_orderid);
-                    $content = $report->data->content;
-                } catch (\Exception $e) {
-                    $content = '';
-                }
-                $this->order->report()->create([
-                    'content' => $content
-                ]);
-            });
+
         }
     }
